requirements.txt
scikit-learn
scipy
tensorflow-gpu==2.6.0
tqdm
numpy